requirements.txt
aiosignal==1.3.1
alembic==1.11.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
config==0.5.1
exceptiongroup==1.1.1
fastapi==0.95.1
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
gunicorn